+++ flexmeasures/cli/__init__.py
@@ -8,6 +8,7 @@ def register_at(app: Flask):
             import flexmeasures.cli.jobs
             import flexmeasures.cli.monitor
             import flexmeasures.cli.data_add
+            import flexmeasures.cli.data_show
             import flexmeasures.cli.data_delete
             import flexmeasures.cli.db_ops
             import flexmeasures.cli.testing  # noqa: F401